+++ changelog.md
@@ -16,6 +16,7 @@ __[changes]__
 __[maintenance]__
 
 * Update certificate information for Blender LTS installer.
+* Update certificate information for CCleaner installer.
 * Update certificate information for HeidiSQL installer.
 * Update certificate information for Node.js installer.
 * Update certificate information for SeaMonkey installers, because installers